refactor: Replace test dependencies implementation for the core:test module
