+++ core/network/build.gradle.kts
@@ -22,7 +22,7 @@ android {
             isMinifyEnabled = false
             proguardFiles(
                 getDefaultProguardFile("proguard-android-optimize.txt"),
-                "proguard-rules.pro",
+                "proguard-rules.pro"
             )
         }
     }
@@ -58,9 +58,4 @@ dependencies {
 //    Hilt
     implementation(libs.hilt.android)
     ksp(libs.hilt.compiler)
-
-//    Test
-    testImplementation(libs.test.junit)
-    androidTestImplementation(libs.test.androidx.junit)
-    androidTestImplementation(libs.test.androidx.espresso.core)
 }
