+++ core/testing/src/androidTest/java/com/octopus/edu/kotlin/core/testing/ExampleInstrumentedTest.kt
@@ -1,10 +1,12 @@
-package com.octopus.edu.kotlin.core.network
+package com.octopus.edu.kotlin.core.testing
 
-import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
-import org.junit.Assert.assertEquals
+import androidx.test.ext.junit.runners.AndroidJUnit4
+
 import org.junit.Test
 import org.junit.runner.RunWith
+
+import org.junit.Assert.*
 
 /**
  * Instrumented test, which will execute on an Android device.
@@ -17,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.octopus.edu.kotlin.core.network.test", appContext.packageName)
+        assertEquals("com.octopus.edu.kotlin.core.testing.test", appContext.packageName)
     }
 }
